completed company list view and added detail view jsp skeleton for company, product and user

+++ twicks/src/main/java/in/ac/vit/twicks/datastorage/dao/AbstractDao.java
@@ -21,7 +21,7 @@ public interface AbstractDao<X> {
 
 	public X update(X x);
 
-	public int getCount();
+	public Long getCount();
 
 	public List<X> get(int first, int pageSize, String sortField,
 			String sortOrder, Map<String, String> filters);
